+++ tests/unit/plugins/openstack/test_openstack.py
@@ -17,8 +17,8 @@ from unittest import TestCase
 
 from cibyl.models.ci.base.environment import Environment
 from cibyl.models.ci.base.job import Job
-from cibyl.outputs.cli.ci.systems.base.colored import \
-    ColoredBaseSystemPrinter as JobPrinter
+from cibyl.outputs.cli.ci.systems.jobs.colored import \
+    ColoredJobsSystemPrinter as JobPrinter
 from cibyl.plugins.openstack.deployment import Deployment
 from cibyl.plugins.openstack.utils import translate_topology_string
 from cibyl.utils.colors import ClearText
